Fixed: Major refactoring how completion blocks are called after fetch.
This is to get the same behavior as 'saveChanges' with completion blocks using two parameters, result and status code. They are always called even when status code != 200. Completion blocks and delegate messages are always sent instead of only one of them.
I have added a check if status code is 200 before doing anything in the completion blocks used in Office.